update by testing .

+++ Bsp/inc/bsp_adc.h
@@ -6,7 +6,7 @@
 
 void Get_Fan_ADC_Fun(uint8_t  channel,uint8_t times);
 
-void Get_Motor_ADC_Fun(uint8_t channel,uint8_t times);
+void Get_Motor_ADC_Fun(uint32_t channel,uint8_t times);
 
 
 extern uint16_t ADC_ConvertedValue[ADC_CHANNEL_NUMBER];
